Checking that files are received before deletion

+++ lib/passy_data/synchronization.dart
@@ -1636,7 +1636,14 @@ class Synchronization {
             _handleException('Failed to decode history entry:\n$e\n$s`.');
             return;
           }
-          _passyEntries.fileIndex!.removeFile(key);
+          // Check that data is received successfully before deletion
+          if (message.binaryObjects != null && remoteFsMeta != null) {
+            if (message.binaryObjects!.isNotEmpty) {
+              _passyEntries.fileIndex!.removeFile(key);
+            }
+          } else if (remoteHistoryEntry.status == EntryStatus.removed) {
+            _passyEntries.fileIndex!.removeFile(key);
+          }
           if (message.binaryObjects != null && remoteFsMeta != null) {
             if (message.binaryObjects!.isNotEmpty) {
               // Save file
